+++ src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -26,6 +26,7 @@ internal sealed class AggregatorStore
     internal readonly int NumberOfMetricPoints;
     internal readonly ConcurrentDictionary<Tags, LookupData>? TagsToMetricPointIndexDictionaryDelta;
     internal readonly Func<ExemplarReservoir?>? ExemplarReservoirFactory;
+    internal readonly bool RecordSum;
     internal long DroppedMeasurements;
 
     private const ExemplarFilterType DefaultExemplarFilter = ExemplarFilterType.AlwaysOff;
@@ -79,7 +80,8 @@ internal sealed class AggregatorStore
         AggregationTemporality temporality,
         int cardinalityLimit,
         ExemplarFilterType? exemplarFilter = null,
-        Func<ExemplarReservoir?>? exemplarReservoirFactory = null)
+        Func<ExemplarReservoir?>? exemplarReservoirFactory = null,
+        bool recordSum = true)
     {
         this.name = metricStreamIdentity.InstrumentName;
 
@@ -87,6 +89,7 @@ internal sealed class AggregatorStore
         // This adjustment accounts for overflow attribute and a case where zero tags are provided.
         // Previously, these were included within the original cardinalityLimit, but now they are explicitly added to enhance clarity.
         this.NumberOfMetricPoints = cardinalityLimit + 2;
+        this.RecordSum = recordSum;
 
         this.metricPoints = new MetricPoint[this.NumberOfMetricPoints];
         this.currentMetricPointBatch = new int[this.NumberOfMetricPoints];